refactor: Move ExceptionHandler to service
